requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.8
pipreqs==0.4.10
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
urllib3==1.25.8
yarg==0.1.9
